Set up a basic game loop with score
When the same card (color) has already been clicked in the past the
game displays an alert to inform the user that the game is over and it
resets the score.
Each time a card is clicked and the game has not been lost the cards
are shuffled to confuse the player.

Create colorsData.js to store the information about the different
colors that can be displayed in game with placeholder data that soon
will be replaced.